backend/requirements.txt
Flask
Flask-Cors
PyPDF2
python-docx
python-dotenv
google-cloud-aiplatform
langchain
langchain-community
langchain-google-vertexai
pypdf
python-docx
faiss-cpu
pytesseract
pdf2image
pillow
google-generativeai
